refactor: Move OrgInstructorFactory helper to common/
